feat(ordering): added edit-orders command to list/remove orders #4/#7; refactor(order/ordering): project refactoring
